+++ tests/fixtures/.php-cs-fixer.dist.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+$finder = PhpCsFixer\Finder::create()
+    ->in(__DIR__)
+    ->name('*.php')
+    ->notName('.php-cs-fixer.dist.php');
+
+return (new PhpCsFixer\Config())
+    ->setRiskyAllowed(true)
+    ->setRules([
+        '@PSR12' => true,
+    ])
+    ->setFinder($finder);